# Conflicts:
#	docker/all-in-one/docker-compose.yml

+++ backend/app/Services/Application/Handlers/CapacityAssignment/DeleteCapacityAssignmentHandler.php
@@ -2,6 +2,10 @@
 
 namespace HiEvents\Services\Application\Handlers\CapacityAssignment;
 
+use HiEvents\DomainObjects\Enums\CapacityChangeDirection;
+use HiEvents\Events\CapacityChangedEvent;
+use HiEvents\Exceptions\ResourceNotFoundException;
+use HiEvents\Models\CapacityAssignment;
 use HiEvents\Repository\Interfaces\CapacityAssignmentRepositoryInterface;
 use HiEvents\Repository\Interfaces\ProductRepositoryInterface;
 use Illuminate\Database\DatabaseManager;
@@ -16,8 +20,22 @@ class DeleteCapacityAssignmentHandler
     {
     }
 
+    /**
+     * @throws ResourceNotFoundException
+     */
     public function handle(int $id, int $eventId): void
     {
+        $capacityAssignment = $this->capacityAssignmentRepository->findFirstWhere([
+            'id' => $id,
+            'event_id' => $eventId,
+        ]);
+
+        if ($capacityAssignment === null) {
+            throw new ResourceNotFoundException(__('Capacity assignment not found'));
+        }
+
+        $productIds = CapacityAssignment::find($id)?->products()->pluck('products.id')->toArray() ?? [];
+
         $this->databaseManager->transaction(function () use ($id, $eventId) {
             $this->productRepository->removeCapacityAssignmentFromProducts(
                 capacityAssignmentId: $id,
@@ -28,5 +46,15 @@ class DeleteCapacityAssignmentHandler
                 'event_id' => $eventId,
             ]);
         });
+
+        if ($capacityAssignment->getCapacity() !== null) {
+            foreach ($productIds as $productId) {
+                event(new CapacityChangedEvent(
+                    eventId: $eventId,
+                    direction: CapacityChangeDirection::INCREASED,
+                    productId: $productId,
+                ));
+            }
+        }
     }
 }
